adminEditor/requirements.txt
requests==2.32.3
csscompressor==0.9.5
jsmin==2.1.1